+++ LethalLevelLoader/Plugin.cs
@@ -21,7 +21,7 @@ namespace LethalLevelLoader
     {
         public const string ModGUID = "imabatby.lethallevelloader";
         public const string ModName = "LethalLevelLoader";
-        public const string ModVersion = "1.2.0.0";
+        public const string ModVersion = "1.2.0.1";
 
         internal static Plugin Instance;